# commons/src/main/resources/swagger.yml
swagger: '2.0'

# Best to be edited on https://editor.swagger.io/#

info:
  version: 1.0.0
  title: Nim
  description: An enterprise implementation of the classic Nim game
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT

basePath: /

tags:
- name: Game data
  description: Provides access to all game resources.

schemes:
- http

paths:
  '/game':
    get:
      summary: Retrieve all game instances
      operationId: getGames
      tags:
      - Game data
      responses:
        '200':
          description: Successful
          schema:
            title: List of games
            type: array
            items:
              $ref: "#/definitions/Game"

    post:
      summary: Create a new game instance
      operationId: createGame
      tags:
      - Game data
      parameters:
      - in: body
        name: newGame
        required: true
        description: Data needed to create a new game.
        schema:
          $ref: '#/definitions/NewGame'
      responses:
        '201':
          description: Game created
        '400':
          description: Bad request. Game not created.
  '/game/{id}':
    patch:
      summary: Make a turn
      operationId: play
      tags:
      - Game data
      parameters:
      - name: id
        in: path
        type: integer
        format: int64
        required: true
        description: Id of the game instance which you want to play
      - in: body
        name: turn
        required: true
        description: The turn you want to play
        schema:
          $ref: '#/definitions/Turn'
      responses:
        '200':
          description: Turn played
          schema:
            $ref: '#/definitions/Game'
        '400':
          description: Invalid turn


definitions:
  Game:
    type: object
    properties:
      id:
        type: integer
        format: int64
      remainingItems:
        type: integer
      initialItems:
        type: integer
      finished:
        type: boolean
      nextPlayer:
        type: string
        enum:
        - HUMAN
        - COMPUTER
      winner:
        type: string
        enum:
        - HUMAN
        - COMPUTER
  NewGame:
    type: object
    properties:
      initialItems:
        type: integer
    required:
    - initialItems
    example:
      initialItems: 13

  Turn:
    type: object
    properties:
      takenItems:
        type: integer
    required:
    - takenItems
    example:
      takenItems: 3